move authAPI to separate file
